New Shell
Changed the pom to use the 3.0 release build of plexus-utils. The
CommandLineService now uses a custom HBCShell. This is a modification
of the BourneShell but this won't escape our pipes anymore. I still
can't figure out why Martin's original code did not suffer from this
problem. My only guess is maybe this shell changed to escape pipes at
sometime, but I can't find the change. And it doesn't look like Martin
ever ran into this problem.